правка замечаний

+++ spring-07/src/main/java/ru/otus/spring/dao/AuthorDao.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 public interface AuthorDao {
     
-    Author getById(int id);
+    Author getById(long id);
     
     List<Author> getAll();
 
-    void deleteById(int id);
+    void deleteById(long id);
 
     void create(Author author);
 }
